added a sweet alert for Are you Sure on delete feedback

+++ src/components/Admin/Admin.jsx
@@ -2,6 +2,7 @@
 import axios from 'axios';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 function Admin() {
 
@@ -31,22 +32,36 @@ function Admin() {
 
     //delete route
     const deleteFeedBack = (event) => {
-        //basically gets the data-id property that was added to the table tr that the button that initiated this function call belongs to
-        //button.    td.         tr
-        const id = event.target.parentElement.parentElement.dataset.id;
-        // console.log(tr.dataset.id);
-        axios({
-            method: 'DELETE',
-            url: `/feedbacks/${id}`,
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                //basically gets the data-id property that was added to the table tr that the button that initiated this function call belongs to
+                //button.    td.         tr
+                const id = event.target.parentElement.parentElement.dataset.id;
+                // console.log(tr.dataset.id);
+                axios({
+                    method: 'DELETE',
+                    url: `/feedbacks/${id}`,
+                })
+                    .then((response) => {
+                        console.log(response.data);
+                        //one was deleted get data again
+                        getFeedBacks(response.data);
+                    })
+                    .catch((error) => {
+                        console.error('DELETE /feedbacks is broken 😢', error);
+                    });
+            }
         })
-            .then((response) => {
-                console.log(response.data);
-                //one was deleted get data again
-                getFeedBacks(response.data);
-            })
-            .catch((error) => {
-                console.error('DELETE /feedbacks is broken 😢', error);
-            });
+
 
     };
